# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "E-commerce API",
    "description": "<h2>To be able to bypass the JWT authentication for the endpoints needing authorization,         login first, then use the token recieved, press the <b>Authorize</b> button and type 'Bearer ' and paste the token copied.         After, you can access the enpoints requiring authorization without filling the <b>Authorization</b> field.</h2>"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
  "paths": {
...
    "/roles/": {
      "get": {
        "tags": [
          "The Role Table Endpoints"
        ],
        "description": "<h3>Gets the role list.</h3>",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "post": {
        "tags": [
          "The Role Table Endpoints"
        ],
        "description": "<h3>If user is authorized and admin, creates a role based on validated parameters provided in the request's body.</h3>",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/roleCreate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "The Role Table Endpoints"
        ],
        "description": "<h3>If user is authorized and admin, changes a role based on validated parameters provided in the request's body. This only happens if it is not a default role.</h3>",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/roleChange"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "The Role Table Endpoints"
        ],
        "description": "<h3>If user is authorized and admin, removes a role based on validated parameters provided in the request's body. This happens only if it is not a default role or no users have it.</h3>",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/roleDelete"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/products/": {
      "get": {
        "tags": [
          "The Product Table Endpoints"
        ],
        "description": "<h3>Gets the product list.</h3>",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "post": {
        "tags": [
          "The Product Table Endpoints"
        ],
        "description": "<h3>If user is authorized and admin, creates a product based on validated parameters provided in the request's body.</h3>",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/productCreate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "The Product Table Endpoints"
        ],
        "description": "<h3>If user is authorized and admin, changes a product based on validated parameters provided in the request's body. When this happens if product's unit price is changed then all cart items related to this product in users' active carts get updated prices. Or if product is soft deleted or quantity equalled to zero then all cart items related to this product in users' active carts get removed.</h3>",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/productChange"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "The Product Table Endpoints"
        ],
        "description": "<h3>If user is authorized and admin, removes a product based on validated parameters provided in the request's body. When this happens then all cart items related to this product in users' active carts get removed, if this product is present.</h3>",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/productDelete"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/admin/products": {
      "get": {
        "tags": [
          "The Admin Front-End Endpoints"
        ],
        "description": "<h3>Renders the products page if the token owner has an admin role. If any of the <b>product</b>, <b>categories</b> and <b>brands</b> request parameters are provided then renders the product page with a filtered product list. If not admin then returns the login page back to the user.</h3>",
        "produces": [
          "text/html"
        ],
        "parameters": [
          {
            "name": "product",
            "in": "query",
            "type": "string"
          },
          {
            "name": "categories",
            "in": "query",
            "type": "string"
          },
          {
            "name": "brands",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "roleCreate": {
      "type": "object",
      "properties": {
        "role": {
          "type": "string",
          "example": "User"
        }
      },
      "required": [
        "role"
      ]
    },
    "roleChange": {
      "type": "object",
      "properties": {
        "roleid": {
          "type": "number",
          "example": 2
        },
        "role": {
          "type": "string",
          "example": "Admin"
        }
      },
      "required": [
        "roleid",
        "role"
      ]
    },
    "roleDelete": {
      "type": "object",
      "properties": {
        "roleid": {
          "type": "number",
          "example": 2
        }
      },
      "required": [
        "roleid"
      ]
    },
...
    "productCreate": {
      "type": "object",
      "properties": {
        "product": {
          "type": "string",
          "example": "XXX"
        },
        "description": {
          "type": "string",
          "example": "XXX"
        },
        "unitprice": {
          "type": "number",
          "example": 0
        },
        "imgurl": {
          "type": "string",
          "example": "XXX"
        },
        "quantity": {
          "type": "number",
          "example": 0
        },
        "brandid": {
          "type": "number",
          "example": 1
        },
        "categoryid": {
          "type": "number",
          "example": 1
        }
      },
      "required": [
        "product",
        "description",
        "unitprice",
        "imgurl",
        "quantity",
        "brandid",
        "categoryid"
      ]
    },
    "productChange": {
      "type": "object",
      "properties": {
        "productid": {
          "type": "number",
          "example": 1
        },
        "product": {
          "type": "string",
          "example": "XXX"
        },
        "description": {
          "type": "string",
          "example": "XXX"
        },
        "unitprice": {
          "type": "number",
          "example": 0
        },
        "imgurl": {
          "type": "string",
          "example": "XXX"
        },
        "quantity": {
          "type": "number",
          "example": 0
        },
        "isdeleted": {
          "type": "number",
          "example": 1
        },
        "brandid": {
          "type": "number",
          "example": 1
        },
        "categoryid": {
          "type": "number",
          "example": 1
        }
      },
      "required": [
        "productid",
        "product",
        "description",
        "unitprice",
        "imgurl",
        "quantity",
        "isdeleted",
        "brandid",
        "categoryid"
      ]
    },
    "productDelete": {
      "type": "object",
      "properties": {
        "productid": {
          "type": "number",
          "example": 1
        }
      },
      "required": [
        "productid"
      ]
    },
...
  }
}